Refactored for heavy clean up
