+++ src/main/java/org/jabref/gui/maintable/columns/FieldColumn.java
@@ -1,7 +1,11 @@
-package org.jabref.gui.maintable;
+package org.jabref.gui.maintable.columns;
 
 import javafx.beans.value.ObservableValue;
 
+import org.jabref.gui.maintable.BibEntryTableViewModel;
+import org.jabref.gui.maintable.MainTableColumnModel;
+import org.jabref.gui.util.ValueTableCellFactory;
+import org.jabref.gui.util.comparator.NumericFieldComparator;
 import org.jabref.model.entry.field.OrFields;
 
 /**
@@ -17,6 +21,12 @@ public class FieldColumn extends MainTableColumn<String> {
 
         setText(getDisplayName());
         setCellValueFactory(param -> getFieldValue(param.getValue()));
+
+        new ValueTableCellFactory<BibEntryTableViewModel, String>()
+                .withText(text -> text)
+                .install(this);
+        this.setComparator(new NumericFieldComparator());
+        this.setSortable(true);
     }
 
     /**
